ASD Week One GHPages Push

+++ js/main.js
@@ -1,6 +1,6 @@
 //Jason Bentley
-//MiU 1207 
-//Project 3
+//ASD 1209 
+//Project 1
 
 
 
@@ -15,36 +15,40 @@ $(document).bind("pageinit", function(){
 		},
 		submitHandler: function(){
 			var data = arform.serializeArray();
+			localStorage.setItem("arform", data);
 			parseRecipeForm(data);
+			// storeData();
 
 		}
 	});
 	
 	//Get ElementById function
-	function gE(x){
-		var theElement = document.getElementById(x);
-		return theElement;
-	}
+	// function gE(x){
+	// 	var theElement = document.getElementById(x);
+	// 	return theElement;
+	// }
 
-// function toggleControls(n){
-// 		switch(n){
-// 			case "on":
-// 				gE("addRecipeForm").style.display = "none";
-// 				gE("clear").style.display = "inline";
-// 				gE("display").style.display = "none";
-// 				gE("addNew").style.display = "inline";
-// 				break;
-// 			case "off":
-// 				gE("addRecipeForm").style.display = "block";
-// 				gE("clear").style.display = "inline";
-// 				gE("display").style.display = "inline";
-// 				gE("addNew").style.display = "none";
-// 				gE("items").style.display = "none";
-// 				break;
-// 			default:
-// 				return false;
-// 		}
-// 	}
+function toggleControls(n){
+		switch(n){
+			case "on":
+				$("#addRecipeForm").hide();
+				$("#clear").show();
+				$("#display1").hide();
+				$("#addNew").show();
+				break;
+			case "off":
+				
+				// $("#addRecipe").show();
+				$("#addRecipeForm").show();
+				$("#clear").show();
+				$("#display1").show();
+				$("#addNew").hide();
+				$("#items").hide();
+				break;
+			default:
+				return false;
+		}
+	}
 	
 
 	function getCheckboxValues(){
@@ -75,63 +79,65 @@ $(document).bind("pageinit", function(){
 		//Get all of our form field value and store in an object.
 		//Object properties contain array with the form label and input values.
 		var item 			= {};
-			item.recipename	= ["Recipe Name:", gE("recipename").value];
-			item.groups 	= ["Group: ",gE("groups").value];
-			item.rating		= ["Rating: ", gE("rating").value];
-			item.date		= ["Date Added: ", gE("date").value];
+			item.recipename	= ["Recipe Name:", $("#recipename").val()];
+			item.groups 	= ["Group: ",$("#groups").val()];
+			item.rating		= ["Rating: ", $("#rating").val()];
+			item.date		= ["Date Added: ", $("#date").val()];
 			item.checks 	= ["Meal Time: " , tcheckedBoxes];
-			item.directions = ["Directions: ", gE("directions").value];
+			item.directions = ["Directions: ", $("#directions").val()];
 		//Save data into Local Storage: Use Stringify to convert the object to a string.
 		localStorage.setItem(id, JSON.stringify(item));
 		alert("Recipe Saved!");
-
+		window.location.reload();
+		
 	}
 	function getData(){
-		// toggleControls("on");
 
+		// toggleControls("on");
 
 		if(localStorage.length === 0){
 			alert("There are no recipes to display! Default Data has been populated!");
 			autoFillData();			
 		}
-		var makeDiv = document.createElement("div");
-		makeDiv.setAttribute("id", "items");
-		var makeList = document.createElement("ul");
+		var makeDiv = $("<div>");
+		makeDiv.attr("id", "items");
+		var makeList = $("<ul>");
 
-		makeDiv.appendChild(makeList);
-		document.getElementById("displayTarget").appendChild(makeDiv);
-		gE("items").style.display = "block";
+		makeDiv.append(makeList);
+		$("#displayTarget").append(makeDiv);
+		$("#items").show();
 		for(var i=0, len=localStorage.length; i<len;i++){
-			var makeLi = document.createElement("li");
-			makeLi.setAttribute("id", "ele");
-			var linksLi= document.createElement("li");
-			linksLi.setAttribute("id", "dL");
-			makeList.appendChild(makeLi);
+			var makeLi = $("<li>");
+			// makeLi.setAttribute("id", "ele");
+			var linksLi= $("<li>");
+			// linksLi.setAttribute("id", "dL");
+			makeList.append(makeLi);
 			var key = localStorage.key(i);
 			var value = localStorage.getItem(key);
 			var obj = JSON.parse(value);
-			var makeSublist = document.createElement("ul");
-			makeLi.appendChild(makeSublist);
+			var makeSublist = $("<ul>");
+			makeLi.append(makeSublist);
 			getImage(obj.groups[1], makeSublist);
 			for(var n in obj){
-				var makeSubli = document.createElement("li");
-				makeSublist.appendChild(makeSubli);
+				var makeSubli = $("<li>");
+				makeSublist.append(makeSubli);
 				var optSubText = obj[n][0]+" "+obj[n][1];
-				makeSubli.innerHTML = optSubText;
-				makeSublist.appendChild(linksLi);
+				makeSubli.text(optSubText);
+				makeSublist.append(linksLi);
 
 			}
 
 			makeItemLinks(localStorage.key(i), linksLi); //Create edit and delete buttons for each item
 		} 
+
 	}
 	//get image for category
 	function getImage(catName, makeSublist){
-		var imageLi = document.createElement("li");
-		makeSublist.appendChild(imageLi);
-		var newImg = document.createElement("img");
-		var setSrc = newImg.setAttribute("src", "images/"+ catName + ".png");
-		imageLi.appendChild(newImg);
+		var imageLi = $("<li>");
+		makeSublist.append(imageLi);
+		var newImg = $("<img>");
+		var setSrc = newImg.attr("src", "images/"+ catName + ".png");
+		imageLi.append(newImg);
 	}
 	//Auto Populate Local Storage
 	function autoFillData(){
@@ -146,46 +152,48 @@ $(document).bind("pageinit", function(){
 	//Make Item Links
 	//Create edit and delete links for eachstored item when disp
 	function makeItemLinks(key, linksLi){
-		var editLink = document.createElement("a");
-		editLink.href = "#";
-		editLink.key = key;
+		var editLink = $("<a>");
+		editLink.attr("href", "#");
+		editLink.attr("key", key);
 		var editText = "Edit Recipe";
-		editLink.addEventListener("click", editItem);
-		editLink.innerHTML = editText;
-		linksLi.appendChild(editLink);
+		editLink.on("click", editItem);
+		editLink.text(editText);
+		linksLi.append(editLink);
 
 		//add line break
-		var breakTag = document.createElement("br");
+		var breakTag = $("<br>");
 		
-		linksLi.appendChild(breakTag);
+		linksLi.append(breakTag);
 
 
 		//add delete single item link
-		var deleteLink = document.createElement("a");
-		deleteLink.href = "#";
-		deleteLink.key = key;
+		var deleteLink = $("<a>");
+		deleteLink.attr("href", "#");
+		deleteLink.attr("key", key);
 		var deleteText = "Delete Recipe";
-		deleteLink.addEventListener("click", deleteItem);
-		deleteLink.innerHTML = deleteText;
-		linksLi.appendChild(deleteLink);
+		deleteLink.on("click", deleteItem);
+		deleteLink.text(deleteText);
+		linksLi.append(deleteLink);
 
 
 	}
 
-	function editItem(){
+	function editItem(key){
 		//grab data from item in l storage
-		var value = localStorage.getItem(this.key);
+		var value = localStorage.getItem(key);
 		var item = JSON.parse(value);
 
 		//Show form
-		// toggleControls("off");
+		toggleControls("off");
 
 		//Populate form fields w/current lstorage vals
-		gE("groups").value = item.groups[1];
-		gE("recipename").value = item.recipename[1];
-		gE("rating").value = item.rating[1];
-		gE("date").value = item.date[1];
-		gE("directions").value = item.directions[1];
+
+		
+		$("#recipename").val(item.recipename[1]);
+		$("#groups").val(item.groups[1]);
+		$("#rating").val(item.rating[1]);
+		$("#date").val(item.date[1]);
+		$("#directions").val(item.directions[1]);
 		
 		var placeValues = function(){
 			var checkboxes = document.getElementById("addRecipeForm").mealTime;
@@ -199,21 +207,24 @@ $(document).bind("pageinit", function(){
 			//console.log(item.checks);//console log to make sure the correct items have been saved
 		};
 		placeValues();
-		console.log(item.checks);
+		// console.log(item.checks);
 		
-		
+		// alert("yo");
 
 		//remove initial listener from the input 'save recipe' button
-		save.removeEventListener("click", storeData);
+		save.off("click", storeData);
 		//Change submit button value to Edit Button
-		gE("submit").value = "Save Edited Recipe";
-		var editSubmit = gE("submit");
+		$("#submit").attr("value", "Save Edited Recipe");
+		var editSubmit = $("#submit");
 		//save the key value estab in this func as a prpty of the editSubmit event
 		//so we can use that value when we save the data we edited.
-		editSubmit.addEventListener("click", submit);
-		editSubmit.key = this.key;
+		editSubmit.on("click", submit);
+		editSubmit.attr ("key", this.key);
+		
 
 	};
+
+	
 
 	function deleteItem(){
 		var ask = confirm("Are you sure you want to delete this recipe?");
@@ -221,6 +232,7 @@ $(document).bind("pageinit", function(){
 			localStorage.removeItem(this.key);
 			alert("Recipe was deleted!");
 			window.location.reload();
+			
 
 		}else{
 			alert("Recipe was NOT deleted.");
@@ -228,14 +240,21 @@ $(document).bind("pageinit", function(){
 	}
 
 	function clearLocal(){
+		var ask = confirm("Are you sure you want to delete ALL of your recipes?");
 		if(localStorage.length === 0){
-			alert("There is no data to clear.")
+			alert("There is no data to clear.");
 		}else{
-			localStorage.clear();
-			alert("All recipes are deleted!");
-			// window.location.reload();
-			return false;
+			if(ask){
+				localStorage.clear();
+				alert("All recipes are deleted!");
+				
+				return false;
+			}else {
+				alert("No Recipes Deleted!");
+			}
+window.location.reload();
 		}
+		return false;
 	}
 
 	
@@ -252,12 +271,12 @@ $(document).bind("pageinit", function(){
 	
 
 	//Set Link and Submit Click Events
-	var displayLink = gE("display1");
-	displayLink.addEventListener("click", getData);
-	var clearLink = gE("clear");
-	clearLink.addEventListener("click", clearLocal);
-	var save = gE("submit");
-	// save.addEventListener("click", validate);
+	var displayLink = $("#display1");
+	displayLink.on("click", getData);
+	var clearLink = $("#clear");
+	clearLink.on("click", clearLocal);
+	var save = $("#submit");
+	save.on("click", storeData);
